Add files via upload
It's not exclusive to humans; countries, too, may incur debt as a means of managing their economies. For instance, substantial investments in infrastructure are vital for providing citizens with a comfortable standard of living. The World Bank is an organization that extends financial assistance to countries through debt.

This notebook delves into the analysis of international debt data compiled by The World Bank. The dataset encompasses details about the debt amounts (in USD) owed by developing countries across various categories. Throughout our exploration, we aim to answer questions such as:

1. What is the total amount of debt owed by the countries listed in the dataset?
2. Which country holds the highest amount of debt, and what is the corresponding figure?
3. What is the average debt amount owed by countries based on different debt indicators?